Creating an Inverse Demand Function in R
Guide to creating an Inverse Demand Function (step function) in R. ggplot2 does not allow for different appearances of vertical and horizontal line segments if you directly use the geom_step() function. I show a way to get around that, by creating vertical and horizontal segments separately, which then allows for differential formatting to indicate the discontinuities in the graph. I also visualise (by adding points) that in cases of indifference (i.e. no consumer surplus), the consumer buys. 
I start by importing a CSV file with the willingness to pay of several consumers, by increment, for up to 3 increments of a good.